// src/shared/dataprovider/api/model/agentTopCardDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ClientUnderAgentDTO } from './clientUnderAgentDTO';
import { CommsPerClient } from './commsPerClient';

export interface AgentTopCardDTO { 
    AgentId?: number;
    WalletBalance?: number;
    RemittanceBalance?: number;
    TotalCiCommission?: number;
    TotalCoCommission?: number;
    ListOfClients?: Array<ClientUnderAgentDTO>;
    CommsPerClient?: Array<CommsPerClient>;
    TotalClientCount?: number;
}